Rename "aux" to "extra"
Contemporary Windows still honours '80s limitations of early MS-DOS filesystems
so that filename "aux" is reserved. CMRX was using it to host systick-based
timing provider. The directory was renamed to "extra" as this timing provider
is not something everyone absolutely has to use.
