+++ src/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_subdirectory(aux)
+add_subdirectory(extra)
 add_subdirectory(os)
 add_subdirectory(lib)
 #if (TESTING)
